Add the rest of the content to the html file
Below the header is some information with the button and the placeholder for the image.

Below the information container are boxes with text. Than there is an quote container and a sign up box with text and a button.

Also there is a footer at the end of the page