Customize body and the container below the header
Info container text and button were placed in a suitable way. The other side of the container, image was added. Colors and fonts were customized to match the color scheme.

Body had some margin added to both side of the content.

Credit of the images creator was added to README file.