Customize information container with boxes
Created boxes to illustrate images and used flex rules to position content correctly. Fonts, colors, margins and padding were also customized in the container.

A page container was created to centralize all content, and an universal selector was added to reset all values. A link was also added to the button. A link was also added to the button.