Customize "call to the action" container
A container was created for the text and the button to separate them from each other. The color theme and fonts were changed to match the model.

Some of the container class names were changed in the html file. In addition, the background color outside the page was changed to black.